refactor: typing/index.d.ts to types
